rename:#5 진입화면 파일명 변경
