+++ src/utils/validate-input.ts
@@ -1,6 +1,4 @@
-{
-  /* 생년월일 */
-}
+
 /**
  * 년도 유효성 검사
  * @param year - 입력된 년도 값
@@ -28,6 +26,13 @@ export const validateMonth = (month: string): string | true => {
   return true;
 };
 
+/**
+ * 일 유효성 검사
+ * @param day - 입력된 일 값
+ * @param year - 입력된 년도 값
+ * @param month - 입력된 월 값 
+ * @returns - 에러 메세지 또는 true
+ */
 export const validateDay = (
   day: string,
   year: string,
@@ -49,9 +54,12 @@ export const validateDay = (
   return true;
 };
 
-{
-  /* 전화번호 */
-}
+
+/**
+ *  전화번호 유효성 검사
+ * @param value - 입력된 전화번호 값
+ * @returns 포맷팅된 전화번호 문자열
+ */
 export const formatPhoneNumber = (value: string) => {
   const cleaned = value.replace(/\D+/g, ''); // 숫자만 남김
   const match = cleaned.match(/^(\d{3})(\d{0,4})(\d{0,4})$/); // 전화번호 형태로 포맷팅
@@ -66,9 +74,13 @@ export const validatePhoneNumber = (value: string) => {
   return numericValue.length === 11;
 };
 
-{
-  /* MBTI */
-}
+
+/**
+ * MBTI 단일 입력란 
+ * @param value - 입력된 MBTI 값
+ * @param index - 입력된 위치 인덱스 (0~3)
+ * @returns true or false
+ */
 export const validateMbtiInput = (value: string, index: number): boolean => {
   const constraints = [
     /^[EIei]$/, // 첫 번째 input: E, I
@@ -80,6 +92,11 @@ export const validateMbtiInput = (value: string, index: number): boolean => {
   return constraints[index].test(value.toUpperCase());
 };
 
+/**
+ * 전체 MBTI 유효성 검사
+ * @param mbti - 입력된 MBTI 문자열
+ * @returns true or false
+ */
 export const isAllMbtiInputsValid = (mbti: string): boolean => {
   const constraints = [/^[EIei]$/, /^[NSns]$/, /^[FTft]$/, /^[PJpj]$/];
 
@@ -90,10 +107,11 @@ export const isAllMbtiInputsValid = (mbti: string): boolean => {
     .every((char, index) => constraints[index].test(char.toUpperCase()));
 };
 
-{
-  /* 닉네임 */
-}
-
+/**
+ * 닉네임 유효성 검사
+ * @param value - 입력된 닉네임 값
+ * @returns 에러 메세지 또는 null
+ */
 export const validateNickname = (value: string): string | null => {
   const nicknameRegex = /^[가-힣a-zA-Z0-9]+$/;
 
@@ -110,10 +128,11 @@ export const isNicknameValid = (value: string): boolean => {
   return validateNickname(value) === null;
 };
 
-{
-  /* 학번 */
-}
-
+/**
+ * 학번 유효성 검사
+ * @param value - 입력된 학번 값
+ * @returns 에러 메세지 또는 null
+ */
 export const validateStudentId = (value: string): string | null => {
   if (!/^\d*$/.test(value)) {
     return '숫자만 입력 가능합니다.';
